+++ tools/ltsview/mathutils.cpp
@@ -28,11 +28,11 @@ namespace MathUtils
 
   float deg_to_rad(float deg)
   {
-    return deg * PI / 180.0f;
+    return deg * static_cast<float>(PI) / 180.0f;
   }
 
   float rad_to_deg(float rad)
   {
-    return rad * 180.0f / PI;
+    return rad * 180.0f / static_cast<float>(PI);
   }
 }
